requirements.txt
discord.py>=2.6.4
requests>=2.32.5
ollama>=0.6.1
